Rename(Member): entity 폴더명 변경 및 MemberDTO->member로 변경,entity폴더로 이동
